Shit ton of comments...
Whole code has been documented - now it should be easy to read :)

+++ Dish_it/src/main/assets/meals.sql
@@ -1,3 +1,3 @@
 BEGIN TRANSACTION;
-CREATE TABLE MealsTable (_id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT, description TEXT, cuisine TEXT, rating INTEGER, ingredients_id TEXT, date DATE, price DOUBLE);
+CREATE TABLE MealsTable (_id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT, description TEXT, cuisine TEXT, rating FLOAT, ingredients_id TEXT, date DATE, price DOUBLE);
 COMMIT;
